+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,13 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Alexander Agafonov on 01.02.2025.
+//
+
+// Структура вопроса
+private struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+}
