+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,13 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Alexander Agafonov on 01.02.2025.
+//
+
+// Структура вью модели результата
+private struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
